# Improve STTClient API for intuitive voice interface usage