… README.md
… examples/README.md
… examples/cli.py
… examples/example_scripts/check_audio_devices.py
… examples/example_scripts/client_example.py
… examples/example_scripts/debug_openai.py
… examples/example_scripts/openai_only.py
… examples/example_scripts/transcribe.py
… examples/example_scripts/wake_word_example.py